requirements.txt
# Core
fastapi
uvicorn

# Database
sqlalchemy
psycopg2-binary
asyncpg
alembic
redis

# Security
python-jose[cryptography]
passlib[bcrypt]
bcrypt
PyJWT
cryptography

# Email
python-multipart
email-validator

# Environment
pydantic-settings
python-dotenv

# CORS
fastapi-cors

# Utils
black
